+++ template-sample-app/template/src/data/sources/kpiSource.js
