+++ template-sample-app/template/src/data/sources/storesSource.js
